requirements.txt
# Core dependencies for Agentic AI HTML5 Parser
groq>=0.9.0       # Groq API for AI agents
pyyaml>=6.0       # YAML parsing for specifications
pytest>=7.0       # Testing framework
pytest-cov>=4.0   # Code coverage reporting
pydantic>=2.0     # Data validation
python-dotenv>=1.0  # Environment variable loading
requests>=2.30    # HTTP requests
lxml>=4.9         # XML/HTML parsing for comparison
jsonschema>=4.17  # JSON schema validation
gitpython>=3.1    # Git operations
beautifulsoup4>=4.12.0
python-docx>=0.8.11
python-pptx>=0.6.21
jinja2>=3.1.2
